big commit! jump analyzer (WIP) is a feature, that can show exactly what happened during a jump each snapshot. Includes features such as position forcing, drawable hitboxes and path drawing. Other features mostly include new function addons and bug/crash fixes

+++ 1_kej/g_trace.cpp
@@ -132,7 +132,7 @@ char cg::BG_AdvanceTrace(BulletTraceResults* br, BulletFireParams* bpp, float _f
 	}
 	return success;
 }
-void __cdecl cg::CG_TracePoint(const vec3_t maxs, trace_t* trace, const vec3_t start, const vec3_t mins, const vec3_t end, int entityNum, int contentMask, uint8_t unknw0, int8_t traceStaticModels)
+void __cdecl cg::CG_TracePoint(const vec3_t maxs, trace_t* trace, const vec3_t start, const vec3_t mins, const vec3_t end, int entityNum, int contentMask, int unknw0, int traceStaticModels)
 {
 
 	_asm
